Refactor Verification, move verificationColumn to its own file
