+++ src/components/Verification/index.js
@@ -1,0 +1,3 @@
+import { verificationColumn } from "./verificationColumn"
+export { verificationColumn }
+export { default as Verification } from "./Verification"
